Partially move datetime namespace
